Standardisation des noms des variables dans les requetes et dans les modeles (ex: Id -> ID), ajustements des services, correction du date format (dd pour jours du mois, DD pour jours de l'annee)

+++ SquidSquAds/src/main/java/com/squidsquads/form/campaign/request/UpdateRequest.java
@@ -1,5 +1,7 @@
 package com.squidsquads.form.campaign.request;
 
+import java.math.BigDecimal;
+
 public class UpdateRequest {
 
     private Long campaignId;
@@ -8,7 +10,7 @@ public class UpdateRequest {
     private Long[] profileIds;
     private String startDate;
     private String endDate;
-    private float budget;
+    private BigDecimal budget;
     private String imageHor;
     private String imageVer;
     private String imageMob;
@@ -37,7 +39,7 @@ public class UpdateRequest {
         return endDate;
     }
 
-    public float getBudget() {
+    public BigDecimal getBudget() {
         return budget;
     }
